+++ geom.go
@@ -96,8 +96,8 @@ func (e ExtentConstraint) Extent() ExtentConstraint {
 	return e
 }
 
-// RenderInfo describes the allocated space for a [FrameSpec] render pass.
-type RenderInfo struct {
+// FrameInfo describes the allocated space for a [FrameSpec] render pass.
+type FrameInfo struct {
 	Width, Height               int     // Total allocated size
 	ContentWidth, ContentHeight int     // Inner content box size
 	FrameWidth, FrameHeight     int     // Total frame size (padding + border + margin)